pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nlb-project"
version = "0.1.0"
description = "Reproducible NLB'21 baseline + improvement project"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "Caleb Telander"}]
dependencies = [
  "numpy>=1.24",
  "pandas==1.3.4",
  "scipy>=1.10",
  "scikit-learn>=1.3",
  "h5py>=3.9",
  "PyYAML>=6.0",
  "nlb-tools==0.0.4",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pip-tools>=7.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-q"
